Quill install

+++ database/migrations/2024_12_18_091321_add_created_by_user_id_to_tasks_table.php
@@ -9,19 +9,23 @@ class AddCreatedByUserIdToTasksTable extends Migration
     public function up()
     {
         Schema::table('tasks', function (Blueprint $table) {
-            $table->foreignId('created_by_user_id')
-                ->nullable()
-                ->after('description')
-                ->constrained('users')
-                ->cascadeOnDelete();
+            if (!Schema::hasColumn('tasks', 'created_by_user_id')) {
+                $table->foreignId('created_by_user_id')
+                    ->nullable()
+                    ->after('description')
+                    ->constrained('users')
+                    ->cascadeOnDelete();
+            }
         });
     }
 
     public function down()
     {
         Schema::table('tasks', function (Blueprint $table) {
-            $table->dropForeign(['created_by_user_id']);
-            $table->dropColumn('created_by_user_id');
+            if (Schema::hasColumn('tasks', 'created_by_user_id')) {
+                $table->dropForeign(['created_by_user_id']);
+                $table->dropColumn('created_by_user_id');
+            }
         });
     }
 }
